LUT26056 : Add javadocs, appointment-ants.properties & move all classes of package pojo in the package web
