+++ src/java/fr/paris/lutece/plugins/appointment/modules/ants/web/PredemandeResponse.java
@@ -32,7 +32,7 @@
  * License 1.0
  */
 
-package fr.paris.lutece.plugins.appointment.modules.ants.pojo;
+package fr.paris.lutece.plugins.appointment.modules.ants.web;
 
 
 import java.util.List;
@@ -41,7 +41,7 @@ import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.fasterxml.jackson.annotation.JsonProperty;
 
 @JsonIgnoreProperties(ignoreUnknown = true)
-public class PredemandePOJO {
+public class PredemandeResponse {
 
 	@JsonProperty("status")
 	private String status;
