+++ webapp/WEB-INF/conf/plugins/appointment-ants.properties
@@ -1,0 +1,9 @@
+appointment-ants.exclude.role=restricted
+
+# example API key, might not work in the future
+ants.api.opt.auth.token=<Insert valid ANTS API Token>
+ants.api.opt.get.status=https://api-coordination.rendezvouspasseport.ants.gouv.fr/api/status?
+ants.display.fieldsErrorMessage=Un ou plusieurs champs ont \u00e9t\u00e9 saisis incorrectement. Veuillez vous assurer de respecter le sch\u00e9ma d'\u00e9criture requis 
+ants.auth.token=x-rdv-opt-auth-token
+ants.session.attribute.name=APPOINTMENT_CODE_PREDEMANDE
+ants.ids_application.parameters=application_ids
